refactor(_stats): split __init__ into per-topic modules (#203)
Mechanical extraction of factrix/_stats/__init__.py (589 lines,
single namespace) into per-topic submodules:

- core.py        — t-stat / p-value / marker / binomial primitives
- hac.py         — Newey-West + Hansen-Hodrick SE / t-test, _resolve_nw_lags
- ols.py         — OLS slope-only + multivariate with NW HAC covariance
- wald.py        — Wald chi-squared test for linear restrictions
- unit_root.py   — Augmented Dickey-Fuller + MacKinnon p-value interp
- diagnostics.py — Ljung-Box portmanteau

__init__.py becomes an index docstring + re-export of all 21 private
symbols, so every existing 'from factrix._stats import X' continues to
resolve unchanged. constants.py untouched. No public surface change
(factrix.stats unaffected). No logic / signature change.

Pre-req for #153 (multiple-testing + bootstrap submodules) — once the
single file would have crossed ~1000 lines.

Refs #203
